char_sprites, chars
In an attempt to build off my Extra class with animated objects I made char_sprites class that gets its information from chars about where to pull the images off the sprite sheet. Issues included attribute error to 'rect' object of charSprite, I didn't write in code that would allow it to animate in place over time, currently that's what I'm working on. The goal is to have adam rise out of the ground at a certain location, so having it be able to determine the image is import. Might have to separate Adam images and have them blit over time to fix this.